Write Accounts mapper with tests
We need a way to have the accounts be mapped to their transactions,
after parsing. That will allow us to make the plots directly with the
same accounts. Instead of passing it around, we can use something like
`Lazy`, built for the first time once we parse it, and we hold this nice
"cache" that we can make computations with.