Start work on Time series plot
Some interesting questions rose: should we work with postive value for
specific accounts? Should we assume the transaction should be the sign
the most amount of transaction have? Should all transactions be
positive?

Hard answers for accounts like "Income:Amazon", since it's used to
add money to the checking account, for example, which will end up
positive. Since ledger is zero sum, Income:Amazon will end up negative,
and most of the values if not all are negative.